requirements.txt
anyio==3.7.0
click==8.1.3
exceptiongroup==1.1.1
fastapi==0.95.2
h11==0.14.0
httptools==0.5.0
idna==3.4
psycopg2-binary==2.9.6
pydantic==1.10.8
python-dotenv==1.0.0
PyYAML==6.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.6.2
uvicorn==0.22.0
uvloop==0.17.0
watchfiles==0.19.0
websockets==11.0.3
